+++ oxidation/libparsec/crates/crypto/src/common/sequester.rs
@@ -1,22 +1,8 @@
 // Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 (eventually AGPL-3.0) 2016-present Scille SAS
 
-use ed25519_dalek::Verifier;
-use rsa::{
-    pkcs8::{
-        der::zeroize::Zeroizing, DecodePrivateKey, DecodePublicKey, EncodePrivateKey,
-        EncodePublicKey,
-    },
-    pss::{Signature, SigningKey, VerifyingKey},
-    signature::RandomizedSigner,
-    PaddingScheme, PublicKey, RsaPrivateKey, RsaPublicKey,
-};
-use serde::{Deserialize, Serialize};
-use serde_bytes::Bytes;
-use sha2::Sha256;
+use crate::{CryptoError, CryptoResult};
 
-use crate::{CryptoError, CryptoResult, SecretKey};
-
-trait EnforceSerialize {
+pub(crate) trait EnforceSerialize {
     const ALGORITHM: &'static [u8];
     const SIZE_IN_BITS: usize;
     const SIZE_IN_BYTES: usize = Self::SIZE_IN_BITS / 8;
@@ -41,7 +27,7 @@ trait EnforceSerialize {
     }
 }
 
-trait EnforceDeserialize {
+pub(crate) trait EnforceDeserialize {
     const ALGORITHM: &'static [u8];
     const SIZE_IN_BITS: usize;
     const SIZE_IN_BYTES: usize = Self::SIZE_IN_BITS / 8;
@@ -63,343 +49,99 @@ trait EnforceDeserialize {
     }
 }
 
-/*
- * PrivateKey
- */
-
-#[derive(Clone)]
-pub struct SequesterPrivateKeyDer(RsaPrivateKey);
-
-crate::impl_key_debug!(SequesterPrivateKeyDer);
-
-impl Default for SequesterPrivateKeyDer {
-    fn default() -> Self {
-        let mut rng = rand_08::thread_rng();
-        Self(RsaPrivateKey::new(&mut rng, Self::SIZE_IN_BITS).expect("Unreachable"))
-    }
-}
-
-impl TryFrom<&[u8]> for SequesterPrivateKeyDer {
-    type Error = CryptoError;
-
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        RsaPrivateKey::from_pkcs8_der(bytes)
-            .map(Self)
-            .map_err(|err| CryptoError::SequesterPrivateKeyDer(err.to_string()))
-    }
-}
-
-impl EnforceDeserialize for SequesterPrivateKeyDer {
-    const SIZE_IN_BITS: usize = Self::SIZE_IN_BITS;
-    const ALGORITHM: &'static [u8] = b"RSAES-OAEP-XSALSA20-POLY1305";
-}
-
-impl SequesterPrivateKeyDer {
-    #[cfg(test)]
-    pub const SIZE_IN_BITS: usize = 1024;
-    #[cfg(not(test))]
-    pub const SIZE_IN_BITS: usize = 4096;
-
-    pub fn dump(&self) -> Zeroizing<Vec<u8>> {
-        self.0.to_pkcs8_der().expect("Unreachable").to_bytes()
-    }
-
-    pub fn dump_pem(&self) -> Zeroizing<String> {
-        self.0
-            .to_pkcs8_pem(rsa::pkcs8::LineEnding::default())
-            .expect("Unreachable")
-    }
-
-    pub fn load_pem(s: &str) -> CryptoResult<Self> {
-        RsaPrivateKey::from_pkcs8_pem(s)
-            .map(Self)
-            .map_err(|err| CryptoError::SequesterPrivateKeyDer(err.to_string()))
-    }
-
-    pub fn decrypt(&self, data: &[u8]) -> CryptoResult<Vec<u8>> {
-        let (cipherkey, ciphertext) = Self::deserialize(data)?;
-        let padding = PaddingScheme::new_oaep::<Sha256>();
-
-        let clearkey = SecretKey::try_from(
-            &self
-                .0
-                .decrypt(padding, cipherkey)
-                .map_err(|_| CryptoError::Decryption)?[..],
-        )?;
-
-        clearkey.decrypt(ciphertext)
-    }
-}
-
-/*
- * PublicKey
- */
-
-#[derive(Clone, PartialEq, Eq, Deserialize)]
-#[serde(try_from = "&Bytes")]
-pub struct SequesterPublicKeyDer(RsaPublicKey);
-
-crate::impl_key_debug!(SequesterPublicKeyDer);
-
-impl From<SequesterPrivateKeyDer> for SequesterPublicKeyDer {
-    fn from(priv_key: SequesterPrivateKeyDer) -> Self {
-        Self(RsaPublicKey::from(priv_key.0))
-    }
-}
-
-impl EnforceSerialize for SequesterPublicKeyDer {
-    const SIZE_IN_BITS: usize = Self::SIZE_IN_BITS;
-    const ALGORITHM: &'static [u8] = b"RSAES-OAEP-XSALSA20-POLY1305";
-}
-
-impl SequesterPublicKeyDer {
-    #[cfg(test)]
-    pub const SIZE_IN_BITS: usize = 1024;
-    #[cfg(not(test))]
-    pub const SIZE_IN_BITS: usize = 4096;
-
-    pub fn dump(&self) -> Vec<u8> {
-        self.0.to_public_key_der().expect("Unreachable").into_vec()
-    }
-
-    pub fn dump_pem(&self) -> String {
-        self.0
-            .to_public_key_pem(rsa::pkcs8::LineEnding::default())
-            .expect("Unreachable")
-    }
-
-    pub fn load_pem(s: &str) -> CryptoResult<Self> {
-        RsaPublicKey::from_public_key_pem(s)
-            .map(Self)
-            .map_err(|err| CryptoError::SequesterPublicKeyDer(err.to_string()))
-    }
-
-    // Encryption format:
-    //   <algorithm name>:<encrypted secret key with RSA key><encrypted data with secret key>
-    pub fn encrypt(&self, data: &[u8]) -> Vec<u8> {
-        let mut rng = rand_08::thread_rng();
-        let padding = PaddingScheme::new_oaep::<Sha256>();
-        let secret_key = SecretKey::generate();
-        let secret_key_encrypted = self
-            .0
-            .encrypt(&mut rng, padding, secret_key.as_ref())
-            .expect("Unreachable");
-
-        EnforceSerialize::serialize(self, &secret_key_encrypted, &secret_key.encrypt(data))
-    }
-}
-
-impl TryFrom<&[u8]> for SequesterPublicKeyDer {
-    type Error = CryptoError;
-
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        RsaPublicKey::from_public_key_der(bytes)
-            .map(Self)
-            .map_err(|err| CryptoError::SequesterPublicKeyDer(err.to_string()))
-    }
-}
-
-impl TryFrom<&Bytes> for SequesterPublicKeyDer {
-    type Error = CryptoError;
-
-    fn try_from(data: &Bytes) -> Result<Self, Self::Error> {
-        Self::try_from(data.as_ref())
-    }
-}
-
-impl Serialize for SequesterPublicKeyDer {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_bytes(&self.dump())
-    }
-}
-
-/*
- * SigningKey
- */
-
-#[derive(Clone)]
-pub struct SequesterSigningKeyDer(SigningKey<Sha256>);
-
-crate::impl_key_debug!(SequesterSigningKeyDer);
-
-impl EnforceSerialize for SequesterSigningKeyDer {
-    const SIZE_IN_BITS: usize = Self::SIZE_IN_BITS;
-    const ALGORITHM: &'static [u8] = b"RSASSA-PSS-SHA256";
-}
-
-impl From<SequesterPrivateKeyDer> for SequesterSigningKeyDer {
-    fn from(priv_key: SequesterPrivateKeyDer) -> Self {
-        Self(SigningKey::from(priv_key.0))
-    }
-}
-
-impl SequesterSigningKeyDer {
-    #[cfg(test)]
-    pub const SIZE_IN_BITS: usize = 1024;
-    #[cfg(not(test))]
-    pub const SIZE_IN_BITS: usize = 4096;
-
-    pub fn dump(&self) -> Zeroizing<Vec<u8>> {
-        self.0.to_pkcs8_der().expect("Unreachable").to_bytes()
-    }
-
-    pub fn dump_pem(&self) -> Zeroizing<String> {
-        self.0
-            .to_pkcs8_pem(rsa::pkcs8::LineEnding::default())
-            .expect("Unreachable")
-    }
-
-    pub fn load_pem(s: &str) -> CryptoResult<Self> {
-        RsaPrivateKey::from_pkcs8_pem(s)
-            .map(|x| Self(SigningKey::from(x)))
-            .map_err(|err| CryptoError::SequesterPrivateKeyDer(err.to_string()))
-    }
-
-    // Signature format:
-    //   <algorithm name>:<signature><data>
-    pub fn sign(&self, data: &[u8]) -> Vec<u8> {
-        let mut rng = rand_08::thread_rng();
-        let signature = self.0.sign_with_rng(&mut rng, data);
-
-        self.serialize(signature.as_ref(), data)
-    }
-}
-
-/*
- * VerifyKey
- */
-
-#[derive(Clone, Deserialize)]
-#[serde(try_from = "&Bytes")]
-pub struct SequesterVerifyKeyDer(VerifyingKey<Sha256>);
-
-crate::impl_key_debug!(SequesterVerifyKeyDer);
-
-impl From<SequesterPublicKeyDer> for SequesterVerifyKeyDer {
-    fn from(pub_key: SequesterPublicKeyDer) -> Self {
-        Self(VerifyingKey::from(pub_key.0))
-    }
-}
-
-impl PartialEq for SequesterVerifyKeyDer {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.as_ref() == other.0.as_ref()
-    }
-}
-
-impl Eq for SequesterVerifyKeyDer {}
-
-impl EnforceDeserialize for SequesterVerifyKeyDer {
-    const SIZE_IN_BITS: usize = Self::SIZE_IN_BITS;
-    const ALGORITHM: &'static [u8] = b"RSASSA-PSS-SHA256";
-}
-
-impl SequesterVerifyKeyDer {
-    #[cfg(test)]
-    pub const SIZE_IN_BITS: usize = 1024;
-    #[cfg(not(test))]
-    pub const SIZE_IN_BITS: usize = 4096;
-
-    pub fn dump(&self) -> Vec<u8> {
-        self.0.to_public_key_der().expect("Unreachable").into_vec()
-    }
-
-    pub fn dump_pem(&self) -> String {
-        self.0
-            .to_public_key_pem(rsa::pkcs8::LineEnding::default())
-            .expect("Unreachable")
-    }
-
-    pub fn load_pem(s: &str) -> CryptoResult<Self> {
-        RsaPublicKey::from_public_key_pem(s)
-            .map(|x| Self(VerifyingKey::from(x)))
-            .map_err(|err| CryptoError::SequesterPublicKeyDer(err.to_string()))
-    }
-
-    pub fn verify(&self, data: &[u8]) -> CryptoResult<Vec<u8>> {
-        let (signature, data) = <Self as EnforceDeserialize>::deserialize(data)?;
-
-        // TODO: It Seems to be a mistake from RustCrypto/RSA
-        // Why should we allocate there ?
-        self.0
-            .verify(data, &Signature::from(signature.to_vec()))
-            .map_err(|_| CryptoError::SignatureVerification)?;
-
-        Ok(data.to_vec())
-    }
-}
-
-impl TryFrom<&[u8]> for SequesterVerifyKeyDer {
-    type Error = CryptoError;
-
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        RsaPublicKey::from_public_key_der(bytes)
-            .map(|x| Self(VerifyingKey::from(x)))
-            .map_err(|err| CryptoError::SequesterPublicKeyDer(err.to_string()))
-    }
-}
-
-impl TryFrom<&Bytes> for SequesterVerifyKeyDer {
-    type Error = CryptoError;
-
-    fn try_from(data: &Bytes) -> Result<Self, Self::Error> {
-        Self::try_from(data.as_ref())
-    }
-}
-
-impl Serialize for SequesterVerifyKeyDer {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_bytes(&self.dump())
-    }
-}
-
 #[test]
-fn test_keys() {
+fn test_key_equality() {
     use hex_literal::hex;
-    assert_eq!(SequesterPrivateKeyDer::SIZE_IN_BITS, 1024);
 
-    let priv_key = SequesterPrivateKeyDer::try_from(
+    let pub_key = SequesterPublicKeyDer::try_from(
         &hex!(
-            "30820276020100300d06092a864886f70d0101010500048202603082025c0201"
-            "0002818100a0b7653b73670b364cd39727a3843502d1ba0402e38f197d4b6ab2"
-            "ea31da8b23cfef82fdc7bf723d0ae6093767aa92274dde3ab0f959f0624d4d5f"
-            "4c7601a561a646e2f323e60f69e52f53158654f0a592ba0f19cebb3f7b2cbc6b"
-            "7920670c6fe1b23c13a77597dc50d34dc1c454c25ee3348cfae7e089bcf50a26"
-            "6b4033f5f502030100010281804c2b1edb32325fe9f8373fa815a30eccab4111"
-            "ad9cd3b12ce49548fe1d6a8a0f0af964878a277da8d8857550c0dce22fa683f0"
-            "24f7c85c58fa71f4fc73e10bd39f8129a2508870f07715c7945038766236face"
-            "3b591bbb5ac4311c807c063ff43b50997152b54a54a4d3fb6ea87ba802ef3831"
-            "c8029642072c045785d9debb01024100ced120f598dde6472aa257d71d7dd7ee"
-            "cfb392694f510babe42620d79374d609e4917f980fca8a5905e8cccdca107364"
-            "9c1bc327c05a7ab271645e12cced3445024100c6efb2dcb8bc84d8ae6ecb17b4"
-            "cb0aa5b4b5685cf60afc560ff2c3268a7be38faabb9da43555b05c6e05b3570c"
-            "69d290fb9b340c275f83794de9bbff4ac94df1024039588b870e08195e0a5851"
-            "7af8567895634a2b82bfd77d210076020d4479d50f912d36eff710f623911be8"
-            "0df7c56ff9a9bf98f160c8b5d4dcd433b18ad90af102403b493609b7785f32e0"
-            "111eaf6aeed3b67c7b4fa5dca17b7ffe72bf9bddcb7c0ed5b7e20c0ce5039118"
-            "2cd4bc8d7380103b1b8ed04c6f9793f017473296cc155102410098b85ecb215a"
-            "60e90d5d8f3e3f70044b47e4e082aa3a5873a5ac0a16ec109d724a44ccb5541e"
-            "8bfe26726bada8c43f344bb953d05fe86afb81c4166f0f2ee425"
+            "30819f300d06092a864886f70d010101050003818d0030818902818100a0b7653b73670b364cd397"
+            "27a3843502d1ba0402e38f197d4b6ab2ea31da8b23cfef82fdc7bf723d0ae6093767aa92274dde3a"
+            "b0f959f0624d4d5f4c7601a561a646e2f323e60f69e52f53158654f0a592ba0f19cebb3f7b2cbc6b"
+            "7920670c6fe1b23c13a77597dc50d34dc1c454c25ee3348cfae7e089bcf50a266b4033f5f5020301"
+            "0001"
         )[..],
     )
     .unwrap();
 
-    let expected_public_key = hex!(
-        "30819f300d06092a864886f70d010101050003818d0030818902818100a0b7653b73670b364cd397"
-        "27a3843502d1ba0402e38f197d4b6ab2ea31da8b23cfef82fdc7bf723d0ae6093767aa92274dde3a"
-        "b0f959f0624d4d5f4c7601a561a646e2f323e60f69e52f53158654f0a592ba0f19cebb3f7b2cbc6b"
-        "7920670c6fe1b23c13a77597dc50d34dc1c454c25ee3348cfae7e089bcf50a266b4033f5f5020301"
-        "0001");
+    let pub_key2 = pub_key.clone();
+
+    assert_eq!(pub_key, pub_key2);
+}
+
+#[cfg(test)]
+mod test {
+    pub use crate::{
+        SequesterPrivateKeyDer, SequesterPublicKeyDer, SequesterSigningKeyDer,
+        SequesterVerifyKeyDer,
+    };
+    pub use hex_literal::hex;
+
+    pub const PRIVATE_KEY_PEM: &str = r#"-----BEGIN PRIVATE KEY-----
+MIICdgIBADANBgkqhkiG9w0BAQEFAASCAmAwggJcAgEAAoGBALLcAKPDtcaJsGnz
+9AxJTSpb4xOxA0+/Hf4O7uDzbPvPYkQAJWzGYNUIR4JzijBF11tYTBlDvATHEj1o
+rAzvJTtO6Neb0J2hkWLcwINmImm3tiyzhYL4owIZBHsIfBG2AYSwST4MHIsdEPnX
+5qLrWv9m9+4YMDGV87zHKrVyB+v9AgMBAAECgYBBfQRO8g3QkAGkCcrF5OD4LYTL
+ZPjNbjDRIS6d9wNkf95+/360gT5bQhjMzvk+C5R6wa27Ym2pYipvia/VXIrIuwoP
+GDKyUg/B2SrDIBgLlleosVmOZwERnY939ShaxXA/TAqT4eXr5uwXm8z/YkledzTV
+iZ2G0ty962SJI9KbEQJBAOsrXVrVvCjK4WVFBcFqKw2Cz6I3+PEPcOXnozMwKKre
+XeT54rPoqfkkoFOPcRngiPPBHxJYMZxZ2gPWN6MkxmMCQQDCs8SRRd/XkwrH9Wga
++kO3sY9pcWNGnHq5psqOEhaOqxozzE5ztT9FCcSAUqMf8onnNXpU+I4o7lQwQACZ
+dmIfAkAmtgiz/yLuBBd+OBJueC+GFdZf+Z6877HB5pNyxaasGdaS7p9mxhHUtTa/
+ComvnMpudYfL2UCxYAkHQKf/7vnJAkEAk1Zqd+zCmWXikLK7Fz8vo4CwoAB4OeUM
+UhVPzvcNLuV4LJ58976+pEXh96GRZAmsJdUoP8jf+0VvXBvy2C7nzQJAGLjFefMr
+vXTMHxDHhuHg4ZJSbJ5BNMZb/HZ5m4KQCt9Gelx/sxZOt3VlCrquCFALxmkeYMhX
+W1pav08hVpEcTg==
+-----END PRIVATE KEY-----
+"#;
+    pub const PUBLIC_KEY_PEM: &str = r#"-----BEGIN PUBLIC KEY-----
+MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQCy3ACjw7XGibBp8/QMSU0qW+MT
+sQNPvx3+Du7g82z7z2JEACVsxmDVCEeCc4owRddbWEwZQ7wExxI9aKwM7yU7TujX
+m9CdoZFi3MCDZiJpt7Yss4WC+KMCGQR7CHwRtgGEsEk+DByLHRD51+ai61r/Zvfu
+GDAxlfO8xyq1cgfr/QIDAQAB
+-----END PUBLIC KEY-----
+"#;
+
+    pub const PRIVATE_KEY_DER: [u8; 634] = hex!(
+        "30820276020100300d06092a864886f70d0101010500048202603082025c02010002818100"
+        "b2dc00a3c3b5c689b069f3f40c494d2a5be313b1034fbf1dfe0eeee0f36cfbcf624400256c"
+        "c660d5084782738a3045d75b584c1943bc04c7123d68ac0cef253b4ee8d79bd09da19162dc"
+        "c083662269b7b62cb38582f8a30219047b087c11b60184b0493e0c1c8b1d10f9d7e6a2eb5a"
+        "ff66f7ee18303195f3bcc72ab57207ebfd0203010001028180417d044ef20dd09001a409ca"
+        "c5e4e0f82d84cb64f8cd6e30d1212e9df703647fde7eff7eb4813e5b4218cccef93e0b947a"
+        "c1adbb626da9622a6f89afd55c8ac8bb0a0f1832b2520fc1d92ac320180b9657a8b1598e67"
+        "01119d8f77f5285ac5703f4c0a93e1e5ebe6ec179bccff62495e7734d5899d86d2dcbdeb64"
+        "8923d29b11024100eb2b5d5ad5bc28cae1654505c16a2b0d82cfa237f8f10f70e5e7a33330"
+        "28aade5de4f9e2b3e8a9f924a0538f7119e088f3c11f1258319c59da03d637a324c6630241"
+        "00c2b3c49145dfd7930ac7f5681afa43b7b18f697163469c7ab9a6ca8e12168eab1a33cc4e"
+        "73b53f4509c48052a31ff289e7357a54f88e28ee543040009976621f024026b608b3ff22ee"
+        "04177e38126e782f8615d65ff99ebcefb1c1e69372c5a6ac19d692ee9f66c611d4b536bf0a"
+        "89af9cca6e7587cbd940b160090740a7ffeef9c902410093566a77ecc29965e290b2bb173f"
+        "2fa380b0a0007839e50c52154fcef70d2ee5782c9e7cf7bebea445e1f7a1916409ac25d528"
+        "3fc8dffb456f5c1bf2d82ee7cd024018b8c579f32bbd74cc1f10c786e1e0e192526c9e4134"
+        "c65bfc76799b82900adf467a5c7fb3164eb775650abaae08500bc6691e60c8575b5a5abf4f"
+        "2156911c4e");
+
+    pub const PUBLIC_KEY_DER: [u8; 162] = hex!(
+        "30819f300d06092a864886f70d010101050003818d0030818902818100b2dc00a3c3b5c689"
+        "b069f3f40c494d2a5be313b1034fbf1dfe0eeee0f36cfbcf624400256cc660d5084782738a"
+        "3045d75b584c1943bc04c7123d68ac0cef253b4ee8d79bd09da19162dcc083662269b7b62c"
+        "b38582f8a30219047b087c11b60184b0493e0c1c8b1d10f9d7e6a2eb5aff66f7ee18303195"
+        "f3bcc72ab57207ebfd0203010001"
+    );
+}
+#[cfg(test)]
+use test::*;
+
+#[test]
+fn test_keys() {
+    assert_eq!(SequesterPrivateKeyDer::SIZE_IN_BITS, 1024);
+
+    let priv_key = SequesterPrivateKeyDer::try_from(&PRIVATE_KEY_DER[..]).unwrap();
 
     let pub_key = SequesterPublicKeyDer::from(priv_key.clone());
 
-    assert_eq!(pub_key.dump(), expected_public_key);
+    assert_eq!(pub_key.dump(), PUBLIC_KEY_DER);
 
     let data = b"Hello world";
 
@@ -413,4 +155,70 @@ fn test_keys() {
     let signed = signing_key.sign(data);
 
     assert_eq!(verify_key.verify(&signed).unwrap(), data);
+}
+
+#[test]
+fn test_pub_signature_verification() {
+    let pub_key_verifier = SequesterVerifyKeyDer::try_from(&PUBLIC_KEY_DER[..]).unwrap();
+
+    let signed_data = hex!(
+        "5253415353412d5053532d5348413235363a0afc141b03789ac3f2c69bd1e577e279d4570b"
+        "f3fe387f389fe52c2b4ac08a9cacbd3c5c1b080cb39969cf3ff7a375619b4b5adc4aef2aec"
+        "2800f6ead1d78019f8e37d036880b71e6ba4e89562e14ab2d6b35d0db4db48d818f8d4395f"
+        "8d692be38fcdfa8d526a352bb811393dd987ed5a8257b7583d145099037178456baf3c4865"
+        "6c6c6f20776f726c64"
+    );
+
+    let verification = pub_key_verifier.verify(&signed_data);
+
+    assert!(verification.is_ok());
+}
+
+#[test]
+fn test_priv_decipher_verification() {
+    let priv_key = SequesterPrivateKeyDer::try_from(&PRIVATE_KEY_DER[..]).unwrap();
+
+    let data = priv_key
+        .decrypt(&hex!(
+            "52534145532d4f4145502d5853414c534132302d504f4c59313330353a278b9346743cc609"
+            "258a4a82023059411d78c29aed9cf9893dc36b1f1f0055e3db8fa3624b4e3ced4fa1d3683b"
+            "97cff2694ddbebeb9a59d1533e9b4dba005958f70a1b7b4b54fef420fb200146a73ac0f457"
+            "168e71decb50d98af9332da36b5143e3470f7858a1a43f0f7ffff6e98e2487579d96a3791d"
+            "69d48ba307e9984dad42781a567fa27e74e9ee88fd945736968855588eec48f43faa396464"
+            "d9a5e8cfd5326ea0193ff5732b51423146d683b74870ed"
+        ))
+        .unwrap();
+
+    assert_eq!(data, b"Hello world");
+}
+
+#[test]
+fn test_import_export() {
+    let priv_key_pem = SequesterPrivateKeyDer::load_pem(PRIVATE_KEY_PEM).unwrap();
+    let priv_key_der = SequesterPrivateKeyDer::try_from(&PRIVATE_KEY_DER[..]).unwrap();
+
+    let pub_key_pem = SequesterPublicKeyDer::load_pem(PUBLIC_KEY_PEM).unwrap();
+    let pub_key_der = SequesterPublicKeyDer::try_from(&PUBLIC_KEY_DER[..]).unwrap();
+
+    assert_eq!(priv_key_pem, priv_key_der);
+    assert_eq!(pub_key_pem, pub_key_der);
+
+    // Also test roundtrip
+    assert_eq!(
+        SequesterPrivateKeyDer::load_pem(&priv_key_pem.dump_pem()).unwrap(),
+        priv_key_pem
+    );
+    assert_eq!(
+        SequesterPrivateKeyDer::try_from(&priv_key_der.dump()[..]).unwrap(),
+        priv_key_der
+    );
+
+    assert_eq!(
+        SequesterPublicKeyDer::load_pem(&pub_key_pem.dump_pem()).unwrap(),
+        pub_key_pem
+    );
+    assert_eq!(
+        SequesterPublicKeyDer::try_from(&pub_key_der.dump()[..]).unwrap(),
+        pub_key_der
+    );
 }
